+++ CHANGELOG.md
@@ -5,3 +5,12 @@
 ## [Unreleased]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+
+## [1.0.0]
+### Added
+- Presentation Folder Generator feature
+
+## [1.1.0]
+### Updated
+- Changed old folder structures to new folder structures
+- Description content